tests/requirements.txt
# Зависимости для запуска тестов инфраструктуры (Этап 1)
pytest==8.3.5
psycopg2-binary==2.9.10
redis==5.2.1
boto3==1.37.23
requests==2.32.3
python-dotenv==1.0.1